Fix a cast issue that broke the run time clock on (very) fast CPUs

+++ app/display.c
@@ -509,7 +509,7 @@ void do_tick(int my_cpu)
     if (clks_per_msec > 0) {
         uint64_t current_time = get_tsc();
 
-        int secs  = (current_time - run_start_time) / (1000 * clks_per_msec);
+        int secs = (current_time - run_start_time) / (1000 * (uint64_t)clks_per_msec);
         int mins  = secs / 60; secs %= 60; act_sec = secs;
         int hours = mins / 60; mins %= 60;
         display_run_time(hours, mins, secs);
